final_projects/YouTube Video to Insightful Article & PDF/requirements.txt
# ============================================
# YouTube Video to Insightful Article & PDF
# Dependencies
# ============================================

# --- Core Framework ---
streamlit>=1.30.0

# --- YouTube Transcript ---
youtube-transcript-api>=0.6.1

# --- LLM Integration ---
langchain>=0.2.0
langchain-google-genai>=1.0.0
google-generativeai>=0.5.0

# --- PDF Generation ---
fpdf2>=2.7.8

# --- Environment Management ---
python-dotenv>=1.0.0

# --- Utilities ---
requests>=2.31.0
